UPDATED
Atualização dos atributos e adição @Enumerated
